Terminata funzione per traduzione parola del giorno estratta

+++ src/common/dto/MyMemoryResponseData.java
@@ -2,5 +2,5 @@ package common.dto;
 
 public class MyMemoryResponseData {
 	public String translatedText;
-	private int match;
+	private float match;
 }
